moved health check to shared
