+++ Portal/src/Datahub.Core/Model/Health/Configuration/InfrastructureHealthCheckConfiguration.cs
@@ -1,3 +1,4 @@
+using Datahub.Shared.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
