+++ Portal/src/Datahub.Portal/Pages/Tools/Health/InfrastructureHealthPage.razor
@@ -1,4 +1,4 @@
-@using Datahub.Core.Model.Health
+@using Datahub.Shared.Entities
 @using System.Text.Json
 @using Datahub.Infrastructure.Services.Helpers
 @using Microsoft.Graph;
