+++ Shared/src/Datahub.Shared/Datahub.Shared.projitems
@@ -14,6 +14,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Configuration\AzureDevOpsConfiguration.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Configuration\QueueConstants.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\AppServiceConfiguration.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Entities\InfrastructureHealthCheck.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\PostgresConfiguration.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\PostgresTier.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\TerraformOrganization.cs" />
